requirements.txt
python-dotenv
requests
bs4
isodate
